+++ core/src/main/scala/org/broadinstitute/dsde/rawls/webservice/AdminApiService.scala
@@ -15,6 +15,7 @@ import org.broadinstitute.dsde.rawls.model.WorkspaceJsonSupport._
 import org.broadinstitute.dsde.rawls.model._
 import org.broadinstitute.dsde.rawls.monitor.migration.MultiregionalBucketMigrationJsonSupport._
 import org.broadinstitute.dsde.rawls.openam.UserInfoDirectives
+import org.broadinstitute.dsde.rawls.submissions.SubmissionsService
 import org.broadinstitute.dsde.rawls.user.UserService
 import org.broadinstitute.dsde.rawls.workspace.WorkspaceService
 import spray.json.DefaultJsonProtocol._
@@ -28,6 +29,7 @@ trait AdminApiService extends UserInfoDirectives {
   import org.broadinstitute.dsde.rawls.model.UserAuthJsonSupport._
 
   val workspaceServiceConstructor: RawlsRequestContext => WorkspaceService
+  val submissionsServiceConstructor: RawlsRequestContext => SubmissionsService
   val userServiceConstructor: RawlsRequestContext => UserService
   val bucketMigrationServiceConstructor: RawlsRequestContext => BucketMigrationService
 
@@ -48,7 +50,7 @@ trait AdminApiService extends UserInfoDirectives {
         path("admin" / "submissions") {
           get {
             complete {
-              workspaceServiceConstructor(ctx).adminListAllActiveSubmissions()
+              submissionsServiceConstructor(ctx).adminListAllActiveSubmissions()
             }
           }
         } ~
@@ -56,7 +58,7 @@ trait AdminApiService extends UserInfoDirectives {
           (workspaceNamespace, workspaceName, submissionId) =>
             delete {
               complete {
-                workspaceServiceConstructor(ctx)
+                submissionsServiceConstructor(ctx)
                   .adminAbortSubmission(WorkspaceName(workspaceNamespace, workspaceName), submissionId)
                   .map { count =>
                     if (count == 1) StatusCodes.NoContent -> None
@@ -73,7 +75,7 @@ trait AdminApiService extends UserInfoDirectives {
         path("admin" / "submissions" / "queueStatusByUser") {
           get {
             complete {
-              workspaceServiceConstructor(ctx).adminWorkflowQueueStatusByUser
+              submissionsServiceConstructor(ctx).adminWorkflowQueueStatusByUser
             }
           }
         } ~
