+++ core/src/main/scala/org/broadinstitute/dsde/rawls/webservice/RawlsApiService.scala
@@ -18,18 +18,19 @@ import io.opentelemetry.context.Context
 import io.sentry.Sentry
 import org.broadinstitute.dsde.rawls.RawlsExceptionWithErrorReport
 import org.broadinstitute.dsde.rawls.billing.BillingProjectOrchestrator
-import org.broadinstitute.dsde.rawls.bucketMigration.{BucketMigrationService, BucketMigrationServiceImpl}
+import org.broadinstitute.dsde.rawls.bucketMigration.BucketMigrationService
 import org.broadinstitute.dsde.rawls.dataaccess.{ExecutionServiceCluster, SamDAO}
 import org.broadinstitute.dsde.rawls.entities.EntityService
 import org.broadinstitute.dsde.rawls.entities.exceptions.DataEntityException
 import org.broadinstitute.dsde.rawls.genomics.GenomicsService
 import org.broadinstitute.dsde.rawls.methods.MethodConfigurationService
 import org.broadinstitute.dsde.rawls.metrics.InstrumentationDirectives
-import org.broadinstitute.dsde.rawls.model.{ApplicationVersion, ErrorReport, RawlsRequestContext, UserInfo}
+import org.broadinstitute.dsde.rawls.model.{ApplicationVersion, ErrorReport, RawlsRequestContext}
 import org.broadinstitute.dsde.rawls.openam.StandardUserInfoDirectives
 import org.broadinstitute.dsde.rawls.snapshot.SnapshotService
 import org.broadinstitute.dsde.rawls.spendreporting.SpendReportingService
 import org.broadinstitute.dsde.rawls.status.StatusService
+import org.broadinstitute.dsde.rawls.submissions.SubmissionsService
 import org.broadinstitute.dsde.rawls.user.UserService
 import org.broadinstitute.dsde.rawls.workspace.{MultiCloudWorkspaceService, WorkspaceService}
 import org.broadinstitute.dsde.workbench.oauth2.OpenIDConnectConfiguration
@@ -108,22 +109,11 @@ trait RawlsApiService
     with VersionApiService
     with ServicePerimeterApiService {
 
-  val multiCloudWorkspaceServiceConstructor: RawlsRequestContext => MultiCloudWorkspaceService
-  val workspaceServiceConstructor: RawlsRequestContext => WorkspaceService
-  val entityServiceConstructor: RawlsRequestContext => EntityService
-  val userServiceConstructor: RawlsRequestContext => UserService
   val genomicsServiceConstructor: RawlsRequestContext => GenomicsService
-  val snapshotServiceConstructor: RawlsRequestContext => SnapshotService
-  val spendReportingConstructor: RawlsRequestContext => SpendReportingService
-  val billingProjectOrchestratorConstructor: RawlsRequestContext => BillingProjectOrchestrator
-  val statusServiceConstructor: () => StatusService
-  val executionServiceCluster: ExecutionServiceCluster
-  val appVersion: ApplicationVersion
   val submissionTimeout: FiniteDuration
   val workbenchMetricBaseName: String
   val samDAO: SamDAO
   val openIDConnectConfiguration: OpenIDConnectConfiguration
-  val methodConfigurationServiceConstructor: RawlsRequestContext => MethodConfigurationService
 
   implicit val executionContext: ExecutionContext
   implicit val materializer: Materializer
@@ -222,6 +212,7 @@ class RawlsApiServiceImpl(val multiCloudWorkspaceServiceConstructor: RawlsReques
                           val billingProjectOrchestratorConstructor: RawlsRequestContext => BillingProjectOrchestrator,
                           val bucketMigrationServiceConstructor: RawlsRequestContext => BucketMigrationService,
                           val methodConfigurationServiceConstructor: RawlsRequestContext => MethodConfigurationService,
+                          val submissionsServiceConstructor: RawlsRequestContext => SubmissionsService,
                           val statusServiceConstructor: () => StatusService,
                           val executionServiceCluster: ExecutionServiceCluster,
                           val appVersion: ApplicationVersion,
